Update 3(I think)
Made it functional! Yeah! Now, assigns "names" to the variables when the buttons are clicked, instead of the final price of the item.
      Ex. *largeButton when clicked 
            size = "large"
This way, the numbers area only added in the whole thing when the code reaches the function. The 'subtotal' variable is created. It will store all the price values depending on what button was clicked.
      Ex. *largeButton clicked and 2 toppings button clicked
            subtotal ...

Well, you get the drill